Refactors to make external reuse easier.
