+++ point_viewer_grpc/Cargo.toml
@@ -24,6 +24,7 @@ clap = "2.32.0"
 futures = "0.1.17"
 grpcio = "0.4"
 protobuf = "2.0.0"
+scoped-pool = "^0.1"
 
 [dependencies.point_viewer]
 path = ".."
